after deleting

+++ todo.js
@@ -3,29 +3,27 @@ const mainbody = document.querySelector("#mainBody")
 const deleteListButton = document.querySelector("#deleteList")
 const renameListButton = document.querySelector("#renameListButton")
 const selectListButton = document.querySelector("#selectListButton")
+// const select = document.querySelector("p")
 
+let deleteElement;
 
 newListButton.addEventListener("click", listCreator)
 deleteListButton.addEventListener("click", listDelete)
 
-const list = JSON.parse(localStorage.getItem('todo'))||[]
 // let list =[]
-console.log(list)
-let count = list.length  ? (Number(list[list.length-1].id)) : 0
+// localStorage.setItem("todo", JSON.stringify(list))
+let list = JSON.parse(localStorage.getItem('todo')) || []
 
-if(list.length) listFromLocalStorage(list)
+let count = list.length ? (Number(list[list.length - 1].id)) : 0
+
+if (list.length) listFromLocalStorage(list)
 
 
-function listFromLocalStorage(list){
-  for(let i of list){
+function listFromLocalStorage(list) {
+  for (let i of list) {
     const res = document.createElement("p")
-    // let check=document.createElement("input")
-    // check.type="checkbox"
-    // check.id="check"
-    // check.appendChild(document.createTextNode(i.name))
-    res.id=i.id
-    res.textContent=i.name
-    // check.appendChild(res)
+    res.id = i.id
+    res.textContent = i.name
     mainbody.appendChild(res)
   }
 }
@@ -42,55 +40,54 @@ function listFromLocalStorage(list){
 //     return dom;
 //   }
 
-function listCreator(event){
-    const Name = prompt("Enter List name")
-    // if(Name){
+function listCreator(event) {
+  const listName = prompt("Enter List name")
+  if (listName) {
     const res = document.createElement("p")
-    let check=document.createElement("input")
-    check.type="checkbox"
-    res.appendChild(check)
+    // let check = document.createElement("input")
+    // check.type = "checkbox"
+    // res.appendChild(check)
     count++
     res.id = count
-    res.textContent = Name
+    res.textContent = listName
     mainbody.appendChild(res)
-    list.push({id:res.id, name:Name})
-    console.log(list,"test")
+    list.push({ id: res.id, name: listName })
     localStorage.setItem("todo", JSON.stringify(list))
-    // console.log(localStorage.getItem())
-    console.log(list)
-  // }
+    const selectP = document.querySelectorAll("p")
+    for (let i of Array.from(selectP)) {
+      i.addEventListener("dblclick", renameList)
+      i.addEventListener("click", deleteOnClick)
+    }
+  }
 }
 // if(list)
 // window.addEventListener("beforeunload",storeLocal)
 
-function storeLocal(){
-localStorage.setItem("todo",JSON.stringify(list))
-}
-function listDelete(event){
-    // const element = document.getElementById(event.target.id);
-    // element.parentNode.removeChild(element);
-}
-
-function renameList(event){
-
-}
-
-function selectList(event){
-
-}
-
-window.addEventListener("mousedown",select)
-
-function select(event){
-  if(event.button=2){
-    event.preventDefault()
-    selectList(event)
-  }
-}
-
-// const selectP = document.querySelector("p")
-// selectP.addEventListener('click',deleteOnClick)
-
-// function deleteOnClick(event){
-//   deleteListButton.style.display="bold";
+// function storeLocal() {
+//   localStorage.setItem("todo", JSON.stringify(list))
 // }
+
+
+
+
+function deleteOnClick(event) {
+  deleteListButton.style.display = "block";
+  deleteElement = event.target
+}
+
+function renameList(event) {
+
+  console.log(event.target.id)
+}
+
+function listDelete() {
+  console.log(deleteElement)
+  let element = document.getElementById(deleteElement.id)
+  element.parentNode.removeChild(element)
+  // deleteElement='';
+  console.log(typeof list)
+  list.splice(deleteElement.id - 1 ,1)
+  console.log(list)
+  console.log(typeof list)
+  localStorage.setItem("todo", JSON.stringify(list))
+}
